+++ src/pages/SonarInsights/Editor/CustomImageSideButton.js
