+++ src/pages/SonarInsights/Editor/Editor.js
