+++ src/pages/SonarInsights/Editor/Editor.scss
