+++ src/pages/SonarInsights/InsightEditor/InsightEditor.module.scss
