+++ src/pages/SonarInsights/InsightEditor/InsightEditorBottom.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Button } from '@santiment-network/ui'
 import moment from 'moment'
-import Timer from './Timer'
-import TagSelector from './TagSelector'
-import styles from './InsightsEditor.module.scss'
+import Timer from '../Timer'
+import TagSelector from '../TagSelector'
+import styles from './InsightEditor.module.scss'
 
 const InsightEditorBottom = ({
   defaultTags,
