+++ src/pages/SonarInsights/InsightEditor/InsightEditorImageModal.css
