+++ src/pages/SonarInsights/InsightEditor/InsightEditorImageModal.js
