+++ src/pages/SonarInsights/InsightEditor/InsightEditorTitle.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import AutoresizeTextarea from './AutoresizeTextarea'
-import styles from './InsightsEditor.module.scss'
+import AutoresizeTextarea from '../AutoresizeTextarea'
+import styles from './InsightEditor.module.scss'
 
 const InsightEditorTitle = ({ className = '', ...props }) => {
   return (
