+++ src/pages/SonarInsights/TagSelector.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { graphql } from 'react-apollo'
 import PropTypes from 'prop-types'
 import { ALL_TAGS_QUERY } from './InsightsGQL'
-import Select from './Select'
+import Select from './Select/Select'
 
 class TagSelect extends Component {
   static defaultProps = {
